apps/api/pyproject.toml
[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[project]
name = "onerhythm-api"
version = "0.1.0"
description = "Initial FastAPI service for OneRhythm."
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
  "alembic==1.14.1",
  "fastapi==0.135.1",
  "numpy==2.4.3",
  "pillow==12.1.1",
  "pypdfium2==5.6.0",
  "python-multipart==0.0.22",
  "pypdf==6.8.0",
  "rapidocr-onnxruntime==1.4.4",
  "starlette==0.52.1",
  "sqlalchemy==2.0.39",
  "uvicorn==0.34.0"
]

[tool.setuptools.packages.find]
include = ["app*"]
